requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Fabric3==1.14.post1
Flask==1.1.2
gunicorn==20.0.4
invoke==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
paramiko==2.7.2
pathlib2==2.3.5
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
six==1.15.0
supervisor==4.2.2
Werkzeug==1.0.1
